pyproject.toml
[project]
name = "systematic-review"
version = "0.1.0"
description = "Systematic review tools."
authors = [
    {name = "Kevin Quinn"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.2.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "langchain-unstructured (>=0.1.6,<0.2.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "langgraph (>=0.4.1,<0.5.0)",
    "unstructured[pdf] (>=0.17.2,<0.18.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "grobid-client-python (>=0.0.9,<0.0.10)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
